+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added session class to set and get values in or from session
 - Added form and controller to change the password
 - Added database seeding system with runner and command create a new seeder
+- Added models for Role and UserRole
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2